+++ packages/pie-toolbox/src/code/graphing/tools/polygon/line.jsx
@@ -5,7 +5,7 @@ import { types, gridDraggable } from '../../../plot';
 import { color } from '../../../render-ui';
 import * as utils from '../../utils';
 import classNames from 'classnames';
-import { correct, disabled, incorrect, missing } from '../shared/styles';
+import { correct, disabled, disabledSecondary, incorrect, missing } from '../shared/styles';
 
 class RawLine extends React.Component {
   static propTypes = {
@@ -32,7 +32,7 @@ class RawLine extends React.Component {
         y1={scale.y(from.y)}
         x2={scale.x(to.x)}
         y2={scale.y(to.y)}
-        className={classNames(classes.line, disabled && classes.disabled, className, classes[correctness])}
+        className={classNames(classes.line, disabled && classes.disabledSecondary, className, classes[correctness])}
         {...rest}
       />
     );
@@ -46,11 +46,15 @@ export const Line = withStyles(() => ({
     stroke: 'transparent',
     '&:hover': {
       strokeWidth: 7,
-      stroke: color.defaults.SECONDARY,
+      stroke: color.defaults.BLACK,
     },
   },
   disabled: {
     ...disabled('stroke'),
+    strokeWidth: 2,
+  },
+  disabledSecondary: {
+    ...disabledSecondary('stroke'),
     strokeWidth: 2,
   },
   correct: {...correct('stroke')},