+++ packages/pie-toolbox/src/code/graphing/tools/shared/point/index.jsx
@@ -1,7 +1,7 @@
 import { withStyles } from '@material-ui/core/styles/index';
 import { gridDraggable } from '../../../../plot';
 import * as utils from '../../../utils';
-import { disabled, correct, incorrect, missing } from '../styles';
+import { disabled, correct, incorrect, missing, disabledSecondary } from '../styles';
 import { RawBp } from './base-point';
 import { RawArrow } from './arrow-point';
 import { color } from '../../../../render-ui';
@@ -29,12 +29,17 @@ const styles = () => {
     point: {
       '& circle, & polygon': {
         cursor: 'pointer',
-        fill: color.defaults.SECONDARY,
+        fill: color.defaults.BLACK,
       },
     },
     disabled: {
       '& circle, & polygon': {
         ...disabled(),
+      },
+    },
+    disabledSecondary: {
+      '& circle, & polygon': {
+        ...disabledSecondary(),
       },
     },
     correct: {